LPS-57882 Match ServiceTrackerCollections
