+++ portal-service/src/com/liferay/portal/kernel/registry/ServiceTrackerCustomizerFactory.java
@@ -69,86 +69,4 @@ public class ServiceTrackerCustomizerFactory {
 		};
 	}
 
-	public static <S> ServiceTrackerCustomizer<S, ServiceWithProperties<S>>
-		serviceWithProperties() {
-
-		return new ServiceTrackerCustomizer<S, ServiceWithProperties<S>>() {
-
-			@Override
-			public ServiceWithProperties<S> addingService(
-				final ServiceReference<S> reference) {
-
-				Registry registry = RegistryUtil.getRegistry();
-
-				final S service = registry.getService(reference);
-
-				if (service == null) {
-					return null;
-				}
-
-				try {
-					final Map<String, Object> properties =
-						Collections.unmodifiableMap(
-							createPropertyMap(reference));
-
-					return new ServiceWithProperties<S>() {
-						@Override
-						public Map<String, Object> getProperties() {
-							return properties;
-						}
-
-						@Override
-						public S getService() {
-							return service;
-						}
-					};
-				}
-
-				catch (Throwable t) {
-					registry.ungetService(reference);
-
-					throw t;
-				}
-			}
-
-			@Override
-			public void modifiedService(
-				ServiceReference<S> reference,
-				ServiceWithProperties<S> service) {
-			}
-
-			@Override
-			public void removedService(
-				ServiceReference<S> reference,
-				ServiceWithProperties<S> service) {
-
-				Registry registry = RegistryUtil.getRegistry();
-
-				registry.ungetService(reference);
-			}
-		};
-	}
-
-	public interface ServiceWithProperties<S> {
-		public Map<String, Object> getProperties();
-
-		public S getService();
-
-	}
-
-	private static <S> Map<String, Object> createPropertyMap(
-		final ServiceReference<S> serviceReference) {
-
-		Map<String, Object> properties = new HashMap<>();
-
-		String[] propertyKeys = serviceReference.getPropertyKeys();
-
-		for (String propertyKey : propertyKeys) {
-			properties.put(
-				propertyKey, serviceReference.getProperty(propertyKey));
-		}
-
-		return properties;
-	}
-
 }
